What is a payload?

+++ src/main/java/com/kelco/kamenridercraft/network/payload/AttributeChangePayload.java
@@ -6,13 +6,13 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.resources.ResourceLocation;
 
-public record AttributeChangePayload(int hand, String attributeName, Float valueChange) implements CustomPacketPayload {
+public record AttributeChangePayload(int hand, String attributeName, Double valueChange) implements CustomPacketPayload {
     public static final Type<AttributeChangePayload> TYPE = new Type<>(ResourceLocation.fromNamespaceAndPath("kamenridercraft", "attribute_change"));
 
     public static final StreamCodec<ByteBuf, AttributeChangePayload> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.VAR_INT, AttributeChangePayload::hand,
             ByteBufCodecs.STRING_UTF8, AttributeChangePayload::attributeName,
-            ByteBufCodecs.FLOAT, AttributeChangePayload::valueChange,
+            ByteBufCodecs.DOUBLE, AttributeChangePayload::valueChange,
             AttributeChangePayload::new
     );
